refactor: move table filter logic to custom useTableFilters hook using useMemo
